// src/models/OTP.ts
export interface IOTP {
  validationCode: string
  referenceNo: string
  sendSms: string
  sendSmsLanguage: string
  clientId: string
  pinType?: string
  validationRefNo: string
  token: string | null
  fp: string | null
  dateTime: Date | string
  version: string
  clientType: string
}

// {"validationRefNo":"C4D164581061DFD174D5A5F4AA7A59870E100BE549EF9B795591313F94383B5D5EFD35083846F2B75A6696E7AEBDC754BE21C396D548682A7945B9480EA23B7AD9F917D32367F71BFC71D2CCA29D456A4974AE4228CDF08AECD6F52EC829904D",
// "token":null,
// "referenceNo":"0000000",
// "sendSmsLanguage":"tur",
// "sendSms":"Y",
// "validationCode":"345345",
// "fp":"","clientId":"34704570",
// "dateTime":"2020-04-06T11:32:09.302Z",
// "version":"34",
// "clientType":"1"}

// 1B51EF5558DCCAB1A116EDF5BDCB35BD40C973487EBB4501669CD482C2581763625D85DC660F99D3E883665EED686F7111D1C0714C79F454019F84CFA95258CB6697932EA14F3EDD8A09EBD856598A9B6B5A333D912A3C704BB764DDF6DC38E79A2EF50C36043CB2879FF35D3A9535A7A66A32E79CB31E3F16B9998472191361F6CDF98DE8DE46EBA4652B32C1B37E37FE66303F